Fix left over distance carry on index change
When resampling the points, left over distance is dropped.
This led to a variable number of points included in the
resample instead of the desired number. This caused issues
when comparing with templates since templates and candidate
gestures would have different number of points and thus would
be penalized with the resultant leading to weaker matches.